monitor v0.2
monitor v0.2

+++ src/main/java/com/okami/core/HeartBeatsThread.java
@@ -6,6 +6,7 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.text.DateFormat;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.LinkedList;
@@ -19,17 +20,20 @@ import org.apache.http.client.ClientProtocolException;
 import org.apache.http.client.HttpClient;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.HttpClients;
+import org.apache.http.message.BasicNameValuePair;
 import org.apache.http.util.EntityUtils;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Component;
 
 import com.okami.bean.ConfigBean;
 import com.okami.bean.GlobaVariableBean;
+import com.okami.common.HttpHandler;
 import com.okami.config.DBConfig;
 import com.okami.dao.impl.CacheLogDao;
 import com.okami.dao.impl.MonitorTaskDao;
 import com.okami.entities.CacheLog;
 import com.okami.util.FileUtil;
+import com.okami.util.WebUtil;
 
 /**
  * 心跳模块以及发送消息模块
@@ -47,7 +51,8 @@ public class HeartBeatsThread extends Thread{
 	
 	private String monitorLogPath ;
 	
-	private String cashLogPath ;
+	private boolean sendLogFlag = false ;
+	
 	
 	/**
 	 * 连接状态
@@ -63,8 +68,6 @@ public class HeartBeatsThread extends Thread{
 	
 	private Date nowTime;
 	
-	private Map<String, String> httpHeaders;
-	
 	/**
 	 * 初始化
 	 */
@@ -74,16 +77,13 @@ public class HeartBeatsThread extends Thread{
 		this.configBean = IOC.instance().getClassobj(ConfigBean.class);
 		this.heartBeatsLogPath = configBean.getLogPath()+File.separator +"heartbeats.log";
 		this.monitorLogPath = configBean.getLogPath()+File.separator +"monitor.log";
-		statusFlag = diffFlag = true;
+		this.statusFlag = this.diffFlag = true;
+		this.sendLogFlag = true;
+		
 		lastTime = null;
-		httpHeaders = new HashMap<String, String>();
-		httpHeaders.put("Charsert", "UTF-8");
-		httpHeaders.put("User-Agent", "Mozilla/5.0 (Windows NT 10.0; WOW64; rv:48.0) Gecko/20100101 Firefox/48.0");
-		httpHeaders.put("Accept", "*/*");
-		httpHeaders.put("Accept-Encoding", "gzip, deflate");
 		return true;
 	}
-	
+
 	public Queue<String> getQHeartBeats(){
 		return qHeartBeats;
 	}
@@ -92,11 +92,69 @@ public class HeartBeatsThread extends Thread{
 		this.qHeartBeats = qHeartBeats;
 	}
 	
+	public boolean getSendLogFlag(){
+		return sendLogFlag;
+	}
+
+	public void setSendLogFlag(boolean sendLogFlag){
+		this.sendLogFlag = sendLogFlag;
+	}
+	
+	/**
+	 * 发送心跳
+	 * @data 2017年4月19日
+	 * @return
+	 */
+	public boolean sendHB(){
+		lastTime = nowTime;
+		HttpHandler httpHandler = IOC.instance().getClassobj(HttpHandler.class);
+		String result =  httpHandler.sendHB();
+		if(result==null || result.indexOf("success")<=0){
+			statusFlag = false;	
+		}else{
+			statusFlag = true;
+		}
+		return true;
+	}
+	
+	/**
+	 * 发送消息，并记录到log中
+	 * @data 2017年4月20日
+	 * @param time
+	 * @param type
+	 * @param content
+	 * @return
+	 */
+	public boolean sendMonitorEvent(String time,String type,String content){
+		HttpHandler httpHandler = IOC.instance().getClassobj(HttpHandler.class);
+		String result = httpHandler.sendMonitorEvent(time, type, content);
+		if(result==null || result.indexOf("true")<=0)
+			statusFlag = false;
+		else
+			statusFlag = true;
+		return true;
+	}
+	
+	/**
+	 * 记录日志
+	 * @data 2017年4月20日
+	 * @param message
+	 * @return
+	 */
+	public boolean writelog(String message){
+		System.out.print(message);
+		FileUtil.write(monitorLogPath, message, true);
+		return true;
+	}
+	
 	public void run(){
+		// 初始化
+
+		
 		if(configBean.getRemoteMode()==true){
 			nowTime = new Date();
 			String message = String.format("[%s][信息][连接服务器 (%s:%s) ...]\r\n",DateFormat.getDateTimeInstance().format(nowTime),configBean.getRhost(),configBean.getRport());
-			FileUtil.write(heartBeatsLogPath, message, true);
+			writelog(message);
 		}
 		
 		while(true){
@@ -104,31 +162,34 @@ public class HeartBeatsThread extends Thread{
 				try {
 					nowTime = new Date();
 					
-					// 心跳
+					// 第一次运行
 					if(lastTime==null){
-						// 第一次运行
-						lastTime = nowTime;
-						statusFlag = httpGet("http://"+configBean.getRhost()+":"+configBean.getRport()+"/monitor/hb");
-					}else if(((nowTime.getTime() - lastTime.getTime())/1000) >= configBean.getDelay()){
-						// 相差大于心跳时间
-						lastTime = nowTime;
-						statusFlag = httpGet("http://"+configBean.getRhost()+":"+configBean.getRport()+"/monitor/hb");
+						sendHB();
+						diffFlag = !statusFlag;
+					}
+					// 相差大于心跳时间
+					else if(((nowTime.getTime() - lastTime.getTime())/1000) >= configBean.getDelay()){
+						sendHB();
 					}
 					
 					// 有消息推送过来
 					if(!qHeartBeats.isEmpty()){
 						// 先存入数据库
 						String text = qHeartBeats.poll();
-						if(text!=null){
-							String[] textList =  text.split("\t");
+						String[] textList =  text.split("\t");
+						String message = String.format("[%s][%s:%s][%s][%s]\r\n",textList[0],configBean.getRhost(),configBean.getRport(),textList[1] ,textList[2]);
+						writelog(message);
+						sendMonitorEvent(textList[0],textList[1],textList[2]);
+						
+						// 网络不通则存入数据库
+						if(!statusFlag){
 							try {
-								System.out.print(textList[0]+"\t"+textList[1]+"\t"+textList[2]+"\r\n");
+								CacheLogDao cacheLogDao = new CacheLogDao();
+								cacheLogDao.setDataSource(new DBConfig().dataSource());
 								CacheLog cacheLog = new CacheLog();
 								cacheLog.setTime(textList[0]);
 								cacheLog.setType(textList[1]);;
 								cacheLog.setEvent(textList[2]);
-								CacheLogDao cacheLogDao = new CacheLogDao();
-								cacheLogDao.setDataSource(new DBConfig().dataSource());
 								cacheLogDao.insertCacheLog(cacheLog);
 							} catch (Exception e) {
 								e.printStackTrace();
@@ -141,30 +202,35 @@ public class HeartBeatsThread extends Thread{
 						diffFlag = statusFlag;
 						if(statusFlag){
 							String message = String.format("[%s][信息][连接服务器 (%s:%s) 成功！]\r\n",DateFormat.getDateTimeInstance().format(nowTime),configBean.getRhost(),configBean.getRport());
+							System.out.print(message);
 							FileUtil.write(heartBeatsLogPath, message, true);
 						}else{
 							String message = String.format("[%s][信息][连接服务器 (%s:%s) 失败！]\r\n",DateFormat.getDateTimeInstance().format(nowTime),configBean.getRhost(),configBean.getRport());
+							System.out.print(message);
 							FileUtil.write(heartBeatsLogPath, message, true);
 						}
 					}
 					
 					// 如果有缓存的log 则进行处理
-					if(statusFlag){
+					if(statusFlag&&sendLogFlag){
 						try {
 							CacheLogDao cacheLogDao = new CacheLogDao();
 							cacheLogDao.setDataSource(new DBConfig().dataSource());
 							List<CacheLog> CacheLogs = cacheLogDao.queryCacheLog();	
-							for(int i=0;i<CacheLogs.size();i++){
+							for(CacheLog cacheLog:CacheLogs){
 								//发送
-								CacheLog cacheLog = CacheLogs.get(i);
-								String message = String.format("[%s][%s:%s][%s][%s]\r\n",cacheLog.getTime(),configBean.getRhost(),configBean.getRport(),cacheLog.getType(),cacheLog.getEvent());
-								statusFlag = httpGet("http://"+configBean.getRhost()+":"+configBean.getRport()+"/monitor/hb");
+								sendMonitorEvent(cacheLog.getTime(),cacheLog.getType(),cacheLog.getEvent());
+								
 								if(!statusFlag){
 									break;
 								}
-	
-								FileUtil.write(monitorLogPath, message, true);
+								
 								cacheLogDao.deleteCacheLog(cacheLog);
+								
+								// 到了心跳时间
+								if(((nowTime.getTime() - lastTime.getTime())/1000) >= configBean.getDelay()){
+									break;
+								}
 							}
 						} catch (Exception e) {
 							e.printStackTrace();
@@ -179,27 +245,4 @@ public class HeartBeatsThread extends Thread{
 		
 	}
 
-	public boolean httpGet(String targetUrl){
-		HttpClient client = HttpClients.createDefault();
-		HttpGet get = new HttpGet(targetUrl);
-		for (String key : httpHeaders.keySet()) {
-        	get.setHeader(key, httpHeaders.get(key));
-        }
-        try {
-        	HttpResponse response = client.execute(get);
-        	HttpEntity httpEntity = response.getEntity();
-        	String result = EntityUtils.toString(httpEntity,"utf-8");
-        	if(result.indexOf("success")<=0){
-        		return false;
-        	}
-        	return true;
-		} catch (UnsupportedEncodingException e) {
-//			e.printStackTrace();
-		} catch (ClientProtocolException e) {
-//			e.printStackTrace();
-		} catch (IOException e) {
-//			e.printStackTrace();
-		}
-        return false;
-	}
 }
